\name{example_data}
\docType{data}
\alias{example_data}
\title{A data set for running examples in user manual}
\description{
  This data set is the biological-effects-only data with amplification constant 1.6.
}
\details{
  The data set contains 192 tumor samples (columns) with 3,523 probes (rows). 
}
\format{A data frame with 3,523 rows by 192 columns.}
\references{
	Yilin Wu*, Becky Wing-Yan Yuen*, Yingying Wei, Li-Xuan Qin^. On data normalization for tumor subtyping with microRNA data, Manuscript.
}
\keyword{datasets}